Infra 재정의  - Realtime sim 데이터와 저장된 데이터 비교를 위해 인프라 분리 진행
