+++ RunSimulation.py
@@ -1,14 +1,10 @@
 import os
-from enum import Enum
-
 import traci
 import pandas as pd
-import math
-from collections import defaultdict
-
-from traci import TraCIException
-import time
 from collections import deque
+
+from Infra import SDetector, SStation, SSection
+
 
 class Config_SUMO:
     # SUMO Configuration File
@@ -20,253 +16,11 @@ class Config_SUMO:
 
     sumoBinary = r'C:/Program Files (x86)/Eclipse/Sumo/bin/sumo-gui'
 
-class Direction(Enum):
-    SB = 0
-    NB = 1
-    EB = 2
-    WB = 3
-
-class InputStation(Enum):
-    SB = '000000'
-    NB = '010021'
-    EB = '020018'
-    WB = '030017'
-
-def get_input_station_value(direction: Direction) -> str:
-    # Direction의 name으로 InputStation을 찾아서 value를 반환
-    return InputStation[direction.name].value
-
-# Detector
-class Detector:
-    def __init__(self, id):
-        self.id = id
-        self.aux, self.bound, self.station_id, self.detector_id = self.parse_detector_id(id)
-        self.minInterval = 30
-        self.speed = 0
-        self.volume = 0
-        self.prevVehicles = tuple()
-    def parse_detector_id(self, id):
-        parts = id.split('_')
-        if len(parts) != 2 or not parts[0].startswith("Det"):
-            raise ValueError(f"Invalid detector ID format: {id}")
-        det_info = parts[1]
-        aux = det_info[0]
-        bound = Direction(int(det_info[1]))
-        station_id = det_info[0:6]
-        detector_id = det_info[6:]
-        return aux, bound, station_id, detector_id
-
-    #update detection data by interval
-    def update(self):
-        vehicle_ids = traci.inductionloop.getLastStepVehicleIDs(self.id)
-        #check duplicated vehicles
-        dupvol = 0
-        if self.prevVehicles is not None:
-            for veh in vehicle_ids:
-                if veh in self.prevVehicles:
-                    dupvol += 1
-
-        self.prevVehicles = vehicle_ids
-
-        self.volume = traci.inductionloop.getLastStepVehicleNumber(self.id) - dupvol
-        # if self.id == 'Det_02000000' or self.id == 'Det_02000001' or self.id == 'Det_12002604':
-        #     print("%s -> v : %d" % (self.id, self.volume))
-            #print('--- lsvid : ', vehicle_ids, type(vehicle_ids))
-
-    def getVolume(self):
-        return self.volume
-
-    def getVehicles(self):
-        return self.prevVehicles
-
-class Station:
-    def __init__(self, id, detectors):
-        self.id = id
-        self.dets = detectors
-        self.direction = None if len(self.dets) == 0 else self.dets[0].bound
-        self.volume = 0
-        self.exitVolume = 0
-        self.inputVeh = set()
-        self.exitVeh = set()
-
-
-    def update(self):
-        self.volume = 0
-        self.exitVolume = 0
-        self.inputVeh = set()
-        self.exitVeh = set()
-
-        for det in self.dets:
-            det.update()
-
-            if det.aux == '1':
-                self.exitVolume += det.getVolume()
-                self.exitVeh.update(det.getVehicles())
-            else:
-                self.volume += det.getVolume()
-                self.inputVeh.update(det.getVehicles())
-
-        self.volume = self.volume if self.volume == 0 or self.volume < len(self.inputVeh) else len(self.inputVeh)
-        self.exitVolume = self.exitVolume if self.exitVolume == 0 or self.exitVolume < len(self.exitVeh) else len(self.exitVeh)
-
-        # if self.id == '020000' or self.id == '120026':
-        #     print('station id',self.id,', volume: ',self.getVolume())
-            #print('station id : ', self.id, 'iv: ',self.inputVeh, 'ev: ', self.exitVeh)
-
-    def getVolume(self):
-        return self.volume
-
-    def getExitVolume(self):
-        return self.exitVolume
-
-    def getVehicleData(self):
-        return list(self.inputVeh), list(self.exitVeh)
-
-    def getInputVehIds(self):
-        return self.inputVeh
-
-    def getExitVehIds(self):
-        return self.exitVeh
-
-class Section:
-    def __init__(self, id, stations):
-        self.id = id
-        self.stations = stations
-        self.direction = None if len(self.stations) == 0 else self.stations[0].direction
-        self.section_co2_emission = 0
-        self.section_volume = 0
-        self.traffic_queue = 0
-        self.section_vehicles = set()
-        self.stop_lane = self.StopLane_position()
-        self.in_dilemmaZone=set()
-
-
-    def check_DilemmaZone(self, time, traffic_light_bound, MinGreenTime, MaxGreenTime):
-        if traffic_light_bound == self.id:
-            DilemmaZone_results = []
-            for vehicle in self.section_vehicles:
-                vehicle_distance_list = []
-                vehicle_position = traci.vehicle.getPosition(vehicle)
-                vehicle_x, vehicle_y = vehicle_position
-                # vehicle_distance = math.sqrt((self.stop_x - vehicle_x) ** 2 + (self.stop_y - vehicle_y) ** 2)
-                # for i in range(0, len(self.stop_lane),2):
-                #     stop_lane_x, stop_lane_y = self.stop_lane[i:i+2]
-                #     distance = math.sqrt((stop_lane_x - vehicle_x) ** 2 + (stop_lane_y - vehicle_y) ** 2)
-                #     vehicle_distance_list.append(distance)
-                # vehicle_distance = min(vehicle_distance_list)
-                # vehicle_speed = traci.vehicle.getSpeed(vehicle)
-                # vehicle_type = traci.vehicle.getTypeID(vehicle)
-                # check_value = self.DilemmaZoneControlSignal(time, vehicle_speed, vehicle_distance, vehicle_type, MinGreenTime, MaxGreenTime)
-                for i in range(0, len(self.stop_lane),2):
-                    stop_lane_x, stop_lane_y = self.stop_lane[i:i+2]
-                    distance = math.sqrt((stop_lane_x - vehicle_x) ** 2 + (stop_lane_y - vehicle_y) ** 2)
-                    if distance <= 120:
-                        vehicle_speed = traci.vehicle.getSpeed(vehicle)
-                        vehicle_type = traci.vehicle.getTypeID(vehicle)
-                        check_value = self.DilemmaZoneControlSignal(time, vehicle_speed, distance, vehicle_type, MinGreenTime, MaxGreenTime)
-                        DilemmaZone_results.append(check_value)
-                    else:
-                        pass
-            if "pass" in DilemmaZone_results:
-                return "pass"
-            elif "pass" not in DilemmaZone_results and "yellow" in DilemmaZone_results:
-                return "yellow"
-            else:
-                return "none"
-        else:
-            check_value = "none"
-            return check_value
-
-    def StopLane_position(self):
-        # print(type(self.id))
-        stop_lane = ()
-        last_station = self.stations[-1]
-        for stop_detector in last_station.dets:
-            lane_id = traci.inductionloop.getLaneID(stop_detector.id)
-            lane_shape = traci.lane.getShape(lane_id)
-            stop_lane_position = lane_shape[-1]
-            # stop_x, stop_y = stop_lane_position
-            stop_lane += stop_lane_position
-        return stop_lane
-
-    def DilemmaZoneControlSignal(self, time, s, d, car_type, MinGreenTime, MaxGreenTime):
-        check = "none"
-        s = s*3.6
-        if time >= MinGreenTime:
-            # if time < MaxGreenTime:
-            if MaxGreenTime <= 10:
-                if car_type == "passenger":
-                    T1 = s / 14
-                    D1 = s * T1
-                    if D1 < d:
-                        check="yellow"
-                        return check
-                    else:
-                        # signal extension
-                        print(s, d, car_type, D1)
-                        check = "pass"
-                        return check
-                else:
-                    T2 = s / 9
-                    D2 = s * T2
-                    if D2 < d:
-                        check = "yellow"
-                        return check
-                    else:
-                        # signal extension
-                        print(s, d, car_type, D2)
-                        check = "pass"
-                        return check
-            else:
-                check="yellow"
-                return check
-        else:
-            return check
-
-    def collect_data(self):
-        return self.section_co2_emission, self.section_volume, self.traffic_queue, list(self.section_vehicles)
-
-    def update(self):
-        self.section_co2_emission = 0
-        self.section_volume = 0
-        removal_veh = list()
-        for i, station in enumerate(self.stations):
-            station.update()
-            if i == 0:
-                self.section_volume += station.getVolume()
-                self.section_vehicles.update(station.getInputVehIds())
-
-            #update input station data according to InputStation Setup
-            if station.id == get_input_station_value(self.direction):
-                self.traffic_queue += station.getVolume()
-
-            self.traffic_queue -= station.getExitVolume()
-            self.section_vehicles.difference_update(station.getExitVehIds())
-            # if self.id == '2':
-            #     if station.getExitVolume() > 0:
-            #         print('----exit vol : ',station.getExitVolume())
-
-        for vehicle in self.section_vehicles:
-            try:
-                if traci.vehicle.getCO2Emission(vehicle) >= 0:
-                    self.section_co2_emission += traci.vehicle.getCO2Emission(vehicle) / 1000
-            except TraCIException:
-                print('------------------------disappear -> ',vehicle)
-                #self.section_vehicles.remove(vehicle)
-                removal_veh.append(vehicle)
-
-        self.section_vehicles.difference_update(removal_veh)
-
-        # if self.id == '2':
-        #     print('Sid : ',self.id, ', Queue : ')
-        #     print('---- VehIds : ', self.section_vehicles)
-        #self.collect_data()
-
 class SumoController:
     def __init__(self, config):
         self.config = config
         self.__set_SUMO()
-        self.detectors = [Detector(detector_id) for detector_id in self.__get_detector_ids(self.config)]
+        self.detectors = [SDetector(detector_id) for detector_id in self.__get_detector_ids(self.config)]
         self.stations = {}
         self.sections = {}
         self.section_results = deque()
@@ -275,9 +29,9 @@ class SumoController:
         self.total_co2_emission = 0
         self.total_volume = 0
         self.__get_station()
-        self.station_objects = {station_id: Station(station_id, detectors) for station_id, detectors in self.stations.items()}
+        self.station_objects = {station_id: SStation(station_id, detectors) for station_id, detectors in self.stations.items()}
         self.__get_section()
-        self.section_objects = {section_id: Section(section_id, stations) for section_id, stations in self.sections.items()}
+        self.section_objects = {section_id: SSection(section_id, stations) for section_id, stations in self.sections.items()}
         self.original_phase_durations={}
         self.stepbySec = 1
         self.colDuration = 30 #seconds
@@ -414,7 +168,7 @@ class SumoController:
         append_result = self.section_results.append
 
         for section_id, section in self.section_objects.items():
-            section_co2_emission, section_volume, traffic_queue, section_vehicles = section.collect_data()
+            section_co2_emission, section_volume, traffic_queue = section.collect_data()
 
             #make total volume
             self.total_volume += section_volume
@@ -426,7 +180,6 @@ class SumoController:
                 'Section_CO2_Emission': section_co2_emission,
                 'Section_Volume': section_volume,
                 'traffic_queue': traffic_queue,
-                'Section_Vehicles': section_vehicles,
                 'sectionBound': str(section.direction)
             })
